+++ src/approvals-service.ts
@@ -308,7 +308,6 @@ export async function getFormApprovalFlowInstanceSubmission(
 
 type FormApprovalFlowResponse = {
   forms: FormTypes.Form[]
-  formApprovalFlows: ApprovalTypes.FormApprovalFlow[]
 }
 
 export async function getFormApprovalFlows(