+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Logic in handling scheduling and payment submission events
+
 ## [0.12.4] - 2021-12-02
 
 ### Added